ml-service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
prophet
networkx
psycopg2-binary
sqlalchemy
python-dotenv
requests
bcrypt
